feat: decouple security components
